refactored main.js on frontEnd and server moving routing logic to imports/routes/routes and moving validaorts/routes/routes and user validation to imports/api/users
