+++ .meteor/packages
@@ -21,4 +21,3 @@ insecure@1.0.7                # Allow all DB writes from clients (for prototypin
 static-html
 fourseven:scss
 accounts-password
-jagi:astronomy
