+++ .meteor/versions
@@ -36,12 +36,10 @@ htmljs@1.0.11
 http@1.4.1
 id-map@1.1.0
 insecure@1.0.7
-jagi:astronomy@2.5.8
 launch-screen@1.1.1
 livedata@1.0.18
 localstorage@1.2.0
 logging@1.1.20
-mdg:validation-error@0.5.1
 meteor@1.9.2
 meteor-base@1.4.0
 minifier-css@1.3.1
