+++ package.json
@@ -14,8 +14,7 @@
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
     "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "simpl-schema": "^1.5.3"
+    "react-router-dom": "^4.3.1"
   },
   "meteor": {
     "mainModule": {
